refactored code (simplifying, improving code readability and keys management)

+++ h2o-extensions/target-encoder/src/main/java/ai/h2o/targetencoding/TargetEncoderFrameHelper.java
@@ -12,8 +12,11 @@ import water.fvec.Vec;
 import water.fvec.task.FilterByValueTask;
 import water.fvec.task.IsNotNaTask;
 import water.fvec.task.UniqTask;
+import water.rapids.Rapids;
+import water.rapids.Val;
 import water.rapids.ast.prims.advmath.AstKFold;
 import water.rapids.ast.prims.mungers.AstGroup;
+import water.util.ArrayUtils;
 import water.util.IcedHashMap;
 
 import java.util.HashMap;
@@ -22,8 +25,30 @@ import java.util.Random;
 
 public class TargetEncoderFrameHelper {
 
-  /** @return the expanded with constant vector Frame, for flow-coding */
-  static Frame addCon(Frame fr, String appendedColumnName, long constant ) { fr.add(appendedColumnName, Vec.makeCon(constant, fr.numRows(), Vec.T_NUM)); return fr; }
+  static public Frame rBind(Frame a, Frame b) {
+    if(a == null) {
+      assert b != null;
+      return b;
+    } else {
+      String tree = String.format("(rbind %s %s)", a._key, b._key);
+      return execRapidsAndGetFrame(tree);
+    }
+  }
+
+  private static Frame execRapidsAndGetFrame(String astTree) {
+    Val val = Rapids.exec(astTree);
+    return register(val.getFrame());
+  }
+
+  /** 
+   * expand the frame with constant vector Frame
+   * @return the index of the new vector.
+   **/
+  static int addCon(Frame fr, String appendedColumnName, long constant ) { 
+    Vec constVec = fr.anyVec().makeCon(constant);
+    fr.add(appendedColumnName, constVec); 
+    return fr.numCols() - 1;
+  }
 
   /**
    * @return frame without rows with NAs in `columnIndex` column
@@ -31,7 +56,9 @@ public class TargetEncoderFrameHelper {
   static Frame filterOutNAsInColumn(Frame fr, int columnIndex) {
     Frame oneColumnFrame = new Frame(fr.vec(columnIndex));
     Frame noNaPredicateFrame = new IsNotNaTask().doAll(1, Vec.T_NUM, oneColumnFrame).outputFrame();
-    return selectByPredicate(fr, noNaPredicateFrame);
+    Frame filtered = selectByPredicate(fr, noNaPredicateFrame);
+    noNaPredicateFrame.delete();
+    return filtered;
   }
 
   /**
@@ -50,19 +77,15 @@ public class TargetEncoderFrameHelper {
 
   private static Frame filterByValueBase(Frame fr, int columnIndex, double value, boolean isInverted) {
     Frame predicateFrame = new FilterByValueTask(value, isInverted).doAll(1, Vec.T_NUM, new Frame(fr.vec(columnIndex))).outputFrame();
-    return selectByPredicate(fr, predicateFrame);
+    Frame filtered = selectByPredicate(fr, predicateFrame);
+    predicateFrame.delete();
+    return filtered;
   }
 
   private static Frame selectByPredicate(Frame fr, Frame predicateFrame) {
-    String[] names = fr.names().clone();
-    byte[] types = fr.types().clone();
-    String[][] domains = fr.domains().clone();
-
-    fr.add("predicate", predicateFrame.anyVec());
-    Frame filtered = new Frame.DeepSelect().doAll(types, fr).outputFrame(Key.<Frame>make(), names, domains);
-    predicateFrame.delete();
-    fr.remove("predicate");
-    return filtered;
+    Vec predicate = predicateFrame.anyVec();
+    Vec[] vecs = ArrayUtils.append(fr.vecs(), predicate);
+    return new Frame.DeepSelect().doAll(fr.types(), vecs).outputFrame(Key.make(), fr._names, fr.domains());
   }
 
   /** return a frame with unique values from the specified column */
@@ -105,19 +128,19 @@ public class TargetEncoderFrameHelper {
    * @param name name of the fold column
    * @param nfolds number of folds
    * @param seed
+   * @return the index of the new column
    */
-  public static Frame addKFoldColumn(Frame frame, String name, int nfolds, long seed) {
+  public static int addKFoldColumn(Frame frame, String name, int nfolds, long seed) {
     Vec foldVec = frame.anyVec().makeZero();
     frame.add(name, AstKFold.kfoldColumn(foldVec, nfolds, seed == -1 ? new Random().nextLong() : seed));
-    return frame;
+    return frame.numCols() - 1;
   }
   
-  static EncodingMaps convertEncodingMapFromFrameToMap(Map<String, Frame> encodingMap) {
+  static EncodingMaps convertEncodingMapValues(Map<String, Frame> encodingMap) {
     EncodingMaps convertedEncodingMap = new EncodingMaps();
     Map<String, FrameToTETableTask> tasks = new HashMap<>();
 
     for (Map.Entry<String, Frame> entry : encodingMap.entrySet()) {
-
       Frame encodingsForParticularColumn = entry.getValue();
       FrameToTETableTask task = new FrameToTETableTask().dfork(encodingsForParticularColumn);
       tasks.put(entry.getKey(), task);
@@ -160,6 +183,7 @@ public class TargetEncoderFrameHelper {
   }
 
   public static void encodingMapCleanUp(Map<String, Frame> encodingMap) {
+    if (encodingMap == null) return;
     for (Map.Entry<String, Frame> map : encodingMap.entrySet()) {
       map.getValue().delete();
     }
